+++ src/cpp/microstrain/connections/connection.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <microstrain/embedded_time.hpp>
-#include <microstrain/span.hpp>
+#include <microstrain/array_view.hpp>
 
 #include <stddef.h>
 #include <stdint.h>
@@ -29,15 +29,15 @@ namespace microstrain
         virtual bool sendToDevice(const uint8_t* data, size_t length) = 0;
         virtual bool recvFromDevice(uint8_t* buffer, size_t max_length, unsigned int wait_time_ms, size_t* length_out, EmbeddedTimestamp* timestamp_out) = 0;
 
-        bool sendToDeviceSpan(microstrain::Span<const uint8_t> data) { return sendToDevice(data.data(), data.size()); }
-        bool recvFromDeviceSpan(microstrain::Span<uint8_t> buffer, unsigned int wait_time_ms, size_t* length_out, EmbeddedTimestamp* timestamp_out)
+        bool sendBufferToDevice(microstrain::ConstU8ArrayView data) { return sendToDevice(data.data(), data.size()); }
+        bool recvBufferFromDevice(microstrain::U8ArrayView buffer, unsigned int wait_time_ms, size_t* length_out, EmbeddedTimestamp* timestamp_out)
         {
             if (!recvFromDevice(buffer.data(), buffer.size(), wait_time_ms, length_out, timestamp_out))
                 return false;
 
             return true;
         }
-        bool recvFromDeviceSpanUpdate(microstrain::Span<uint8_t>& buffer, unsigned int wait_time_ms, EmbeddedTimestamp* timestamp_out)
+        bool recvFromDeviceAndUpdateUint8ArrayView(microstrain::U8ArrayView& buffer, unsigned int wait_time_ms, EmbeddedTimestamp* timestamp_out)
         {
             size_t length = 0;
             if (!recvFromDevice(buffer.data(), buffer.size(), wait_time_ms, &length, timestamp_out))