+++ src/cpp/microstrain/serialization/serializer.hpp
@@ -2,7 +2,7 @@
 
 #include "readwrite.hpp"
 
-#include "../span.hpp"
+#include <microstrain/array_view.hpp>
 
 #include <array>
 
@@ -63,7 +63,7 @@ public:
     SerializerBase() = default;
     SerializerBase(uint8_t* ptr, size_t capacity, size_t offset=0) : m_ptr(ptr), m_size(capacity), m_offset(offset) {}
     SerializerBase(const uint8_t* ptr, size_t size, size_t offset=0) : m_ptr(const_cast<uint8_t*>(ptr)), m_size(size), m_offset(offset) {}
-    SerializerBase(microstrain::Span<const uint8_t> buffer, size_t offset=0) : m_ptr(const_cast<uint8_t*>(buffer.data())), m_size(buffer.size()), m_offset(offset) {}
+    SerializerBase(microstrain::ConstU8ArrayView buffer, size_t offset=0) : m_ptr(const_cast<uint8_t*>(buffer.data())), m_size(buffer.size()), m_offset(offset) {}
 
     size_t capacity()   const { return m_size;                 }  ///< Returns the total size of the buffer.
     size_t offset()     const { return m_offset;               }  ///< Returns the current read or write offset.
@@ -497,29 +497,29 @@ size_t extract(Serializer<E>& serializer, T* values, size_t count)
 ///@tparam T Type of array elements. Automatically deduced from the values parameter.
 ///
 ///@param serializer Serializer object pointing to the destination buffer.
-///@param values     Span containing pointer and count.
+///@param values     ArrayView containing pointer and count.
 ///
 ///@returns The total number of bytes written.
 ///
 template<serialization::Endian E, class T>
-size_t insert(Serializer<E>& serializer, microstrain::Span<const T> values)
+size_t insert(Serializer<E>& serializer, microstrain::ArrayView<const T> values)
 {
     return insert(serializer, values.data(), values.size());
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-///@brief Reads an array from a Serializer via a Span.
+///@brief Reads an array from a Serializer via an ArrayView.
 ///
 ///@tparam E Endianness of buffer. Automatically deduced from the serializer parameter.
 ///@tparam T Type of array elements. Automatically deduced from the values parameter.
 ///
 ///@param serializer Serializer object pointing to the source buffer.
-///@param values     Span containing pointer and count.
+///@param values     ArrayView containing pointer and count.
 ///
 ///@returns The total number of bytes read.
 ///
 template<serialization::Endian E, class T>
-size_t extract(Serializer<E>& serializer, microstrain::Span<const T> values)
+size_t extract(Serializer<E>& serializer, microstrain::ConstArrayView<T> values)
 {
     return extract(serializer, values.data(), values.size());
 }
@@ -776,11 +776,11 @@ bool extract(T& value, const uint8_t* buffer, size_t buffer_length, size_t offse
 
 
 //
-// Raw buffer - Span version
+// Raw buffer - ArrayView version
 //
 
 ////////////////////////////////////////////////////////////////////////////////
-///@brief Serializes a value to a raw byte buffer (span version).
+///@brief Serializes a value to a raw byte buffer (ArrayView version).
 ///
 /// Use this overload to write a single value without needing to
 /// manually construct a Serializer object.
@@ -798,13 +798,13 @@ bool extract(T& value, const uint8_t* buffer, size_t buffer_length, size_t offse
 ///@returns True otherwise.
 ///
 template<serialization::Endian E, class T>
-bool insert(T value, microstrain::Span<uint8_t> buffer, size_t offset=0, bool exact_size=false)
+bool insert(T value, microstrain::ArrayView<uint8_t> buffer, size_t offset=0, bool exact_size=false)
 {
     return insert<E,T>(value, buffer.data(), buffer.size(), offset, exact_size);
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-///@brief Deserializes a value from a raw byte buffer (span version).
+///@brief Deserializes a value from a raw byte buffer (ArrayView version).
 ///
 /// Use this overload to read a single value without needing to
 /// manually construct a Serializer object.
@@ -813,7 +813,7 @@ bool insert(T value, microstrain::Span<uint8_t> buffer, size_t offset=0, bool ex
 ///@tparam T Type of value. Automatically deduced from the value parameter.
 ///
 ///@param value         Parameter to deserialize. This can be any serializable type.
-///@param buffer        Source buffer span.
+///@param buffer        Source buffer.
 ///@param offset        Starting offset (default 0).
 ///@param exact_size    Returns true only if exactly buffer.size()-offset bytes are read. Default false.
 ///
@@ -822,7 +822,7 @@ bool insert(T value, microstrain::Span<uint8_t> buffer, size_t offset=0, bool ex
 ///@returns True otherwise.
 ///
 template<serialization::Endian E, class T>
-bool extract(T& value, microstrain::Span<const uint8_t> buffer, size_t offset=0, bool exact_size=false)
+bool extract(T& value, microstrain::ConstU8ArrayView buffer, size_t offset=0, bool exact_size=false)
 {
     return extract<E,T>(value, buffer.data(), buffer.size(), offset, exact_size);
 }
@@ -882,11 +882,11 @@ std::optional<T> extract(const uint8_t* buffer, size_t length, size_t offset, bo
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-///@brief Reads a value from a raw byte span and returns it via std::optional.
+///@brief Reads a value from a raw byte view and returns it via std::optional.
 ///
 /// This overload is only enabled if std::optional is supported.
 ///
-///@see bool extract(T& value, microstrain::Span<const uint8_t> buffer, size_t offset=0, bool exact_size=false)
+///@see bool extract(T& value, microstrain::ConstUint8ArrayView buffer, size_t offset=0, bool exact_size=false)
 ///
 ///@tparam E Endianness of buffer. Must be manually specified.
 ///@tparam T Type of value. Must be manually specified.
@@ -898,7 +898,7 @@ std::optional<T> extract(const uint8_t* buffer, size_t length, size_t offset, bo
 ///@returns The value read from the buffer, or std::nullopt if it couldn't be read.
 ///
 template<class T, serialization::Endian E>
-std::optional<T> extract(microstrain::Span<const uint8_t> buffer, size_t offset, bool exact_size=false)
+std::optional<T> extract(microstrain::ConstU8ArrayView buffer, size_t offset, bool exact_size=false)
 {
     T value;
     if(extract<E,T>(value, buffer.data(), buffer.size(), offset, exact_size))