+++ src/cpp/mip/mip_field.hpp
@@ -3,7 +3,7 @@
 #include "mip_descriptors.hpp"
 #include "mip_serialization.hpp"
 
-#include <microstrain/span.hpp>
+#include <microstrain/array_view.hpp>
 
 #include <mip/mip_field.h>
 #include <mip/mip_offsets.h>
@@ -36,8 +36,8 @@ public:
     /// Creates a %Field class from the mip_field C struct.
     FieldView(const C::mip_field_view& other) { std::memcpy(static_cast<C::mip_field_view*>(this), &other, sizeof(C::mip_field_view)); }
 
-    /// Construct from descriptor and payload (composite span version)
-    FieldView(CompositeDescriptor descriptor, microstrain::Span<const uint8_t> payload) : FieldView(descriptor.descriptorSet, descriptor.fieldDescriptor, payload.data(), uint8_t(payload.size())) {}
+    /// Construct from descriptor and payload bytes
+    FieldView(CompositeDescriptor descriptor, microstrain::ConstU8ArrayView payload) : FieldView(descriptor.descriptorSet, descriptor.fieldDescriptor, payload.data(), uint8_t(payload.size())) {}
 
     //
     // C function wrappers
@@ -54,16 +54,35 @@ public:
     ///@copydoc mip::C::mip_field_total_length
     uint8_t totalLength() const { return C::mip_field_total_length(this); }
     ///@copydoc mip::C::mip_field_payload
-    const uint8_t* payload() const { return C::mip_field_payload(this); }
+    const uint8_t* payloadPointer() const { return C::mip_field_payload(this); }
 
     ///@brief Index the payload at the given location.
-    ///@param index
+    ///@param index Byte index into payload. 0 <= index < payloadLength().
     ///@returns payload byte
-    uint8_t payload(unsigned int index) const { return payload()[index]; }
+    uint8_t payload(size_t index) const { assert(index < payloadLength()); return payloadPointer()[index]; }
 
-    uint8_t operator[](unsigned int index) const { return payload(index); }
+    uint8_t operator[](size_t index) const { return payload(index); }
 
-    microstrain::Span<const uint8_t> payloadSpan() const { return {payload(), payloadLength()}; }
+    ///@brief Get a const view of the payload data.
+    microstrain::ConstU8ArrayView payload() const { return {payloadPointer(), payloadLength()}; }
+
+    ///@brief Gets a view of the entire field, including the header.
+    ///
+    ///@warning FieldView contains a descriptor and payload pointer. It does not
+    ///         guarantee the descriptors are also stored in the same buffer as
+    ///         the payload.
+    ///         This function operates by using a negative index from the
+    ///         payload pointer, so it is up to the user to ensure such an index
+    ///         is safe. Generally, if the field was obtained from an existing
+    ///         packet then the header bytes will be available. E.g. the field
+    ///         was obtained via `PacketView::firstField()`,
+    ///         `for(FieldView field : packet)`, etc. then the whole packet,
+    ///         including all fields and their headers, should exist in the
+    ///         buffer. Standalone FieldViews created from a descriptor and
+    ///         payload array may not have header bytes and this function isn't
+    ///         safe in that case.
+    ///
+    microstrain::ConstU8ArrayView data() const { return {payloadPointer()-C::MIP_HEADER_LENGTH, size_t(payloadLength()+C::MIP_HEADER_LENGTH)}; }
 
     ///@copydoc mip::C::mip_field_is_valid
     bool isValid() const { return C::mip_field_is_valid(this); }
@@ -89,7 +108,7 @@ public:
     ///         too few bytes (or to many if exact_size is specified). The field data is not
     ///         valid unless this function returns true.
     template<class FieldType>
-    bool extract(FieldType& field, bool exact_size=true) const { return microstrain::extract<microstrain::serialization::Endian::big>(field, payload(), payloadLength(), 0, exact_size); }
+    bool extract(FieldType& field, bool exact_size=true) const { return microstrain::extract<microstrain::serialization::Endian::big>(field, payload(), 0, exact_size); }
 
 
     ///@brief Determines if the field holds data (and not a command, reply, or response).