+++ src/cpp/mip/mip_logging.hpp
@@ -41,7 +41,7 @@ namespace mip
 ///@returns True if successful.
 ///@returns False in case of insufficient buffer space.
 ///
-inline bool formatPacketBytes(microstrain::Span<char> buffer, size_t* index, const PacketView& packet)
+inline bool formatPacketBytes(microstrain::CharArrayView buffer, size_t* index, const PacketView& packet)
 {
     return mip::C::mip_format_packet_bytes(buffer.data(), buffer.size(), index, &packet);
 }
@@ -70,7 +70,7 @@ inline bool formatPacketBytes(microstrain::Span<char> buffer, size_t* index, con
 ///@returns True if successful.
 ///@returns False in case of insufficient buffer space.
 ///
-inline bool formatPacket(microstrain::Span<char> buffer, size_t* index, const PacketView& packet)
+inline bool formatPacket(microstrain::CharArrayView buffer, size_t* index, const PacketView& packet)
 {
     return mip::C::mip_format_packet(buffer.data(), buffer.size(), index, &packet);
 }
@@ -101,7 +101,7 @@ inline bool formatPacket(microstrain::Span<char> buffer, size_t* index, const Pa
 ///@returns True if successful.
 ///@returns False in case of insufficient buffer space.
 ///
-inline bool formatField(microstrain::Span<char> buffer, size_t* index, const FieldView& field)
+inline bool formatField(microstrain::CharArrayView buffer, size_t* index, const FieldView& field)
 {
     return mip::C::mip_format_field(buffer.data(), buffer.size(), index, &field);
 }