+++ CHANGELOG.md
@@ -33,6 +33,8 @@ Forthcoming
   * Setup test suite to be easily run locally
 ### Interface Changes
 * MIP parser callbacks must now return bool (return true to preserve existing behavior).
+* microstrain::Span has been renamed to microstrain::ArrayView.
+* Certain PacketView and FieldView accessors have been renamed for clarity and consistency.
 ### Bug Fixes
 * Fixed possible infinite loop in MIP parser with timed-out packets.